NEXUS/requirements.txt
python-docx==1.2.0
lxml>=6.0.0
typing_extensions>=4.9.0
Pillow>=10.0.0
openpyxl>=3.1.0
pandas>=2.0.0
psycopg2-binary>=2.9.0

# Sistema Modular (Novas dependências)
docxcompose>=1.4.0
docx2pdf>=0.1.8

